+++ public/js/graficar.js
@@ -18,8 +18,33 @@ $(document).ready(function() {
     initChartist();
 });
 
-function UpdateChar(data) {
-    char.update(data);
+
+function cargarDatos() {
+
+    var datosConsulta = []
+    var e = document.getElementById("moduloGenerico");
+    var competencia = e.options[e.selectedIndex].value;
+
+    var request = {
+        modulo: competencia,
+        tiempo: "genericas2016"
+    }
+    consultaGenericasPosicion(request).then(function(data) {
+        for (let i = 1; i < data.response.length; i++) {
+            datosConsulta[i] = { meta: data.response[i].INSTITUCION, value: parseInt(data.response[i].PROMEDIO) }
+        }
+        data = {
+            series: [
+                datosConsulta
+            ]
+        };
+
+        char.update(data);
+    })
+}
+
+function actualizarGrafica() {
+    cargarDatos()
 }
 
 function initChartist() {
